test: tarantool -e test add more details and avoid busy loop
Background: the test fails on waiting the process termination, but
surprisingly only on GitHub hosted runners.

My guess is that the process becomes zombie, but we don't collect it.

A busy loop is not good in general. It is better to pass execution to
other fibers if there is nothing to do. This way the waiting will not
eat 100% CPU time.

Added collect more details in tarantool -e 'assert(false)' test.